clean up (part 3): get rid of ImportRewriter and one more class
Removes classes ImportRewriter and ImportSpacerUserPreferenceHelper.
